requirements.txt
PySimpleGUI==4.60.4
Pillow==9.5.0
matplotlib==3.7.1
pandas==1.5.3
numpy==1.24.2
notebook==6.5.3
streamlit==1.23.1
